Completed viewing ideas.
Ajax was added to index.js file to get the idea data, switched from using minitest to rspec, and implemented the
addition of post requests to create new ideas from browser. Test frameworks were changed to issues with capybara and
minitest not able to see javascript. Integration tests work when ran solo, but when ran as a whole with other tests, the
integration test fails. Testing up to this point has been slow and will be put on the back burner until main functionality
is completed.

Next steps is to fully implement Adding a new idea to the database from the web browser via ajax calls.