Completed update ideas.
Added the openEdit, updateIdea, and cancelEdit jQuery functions. One current draw back
is that the modal does not close on clicks, and currently not sure how to fix that. Other than
that there is no un wanted functionality.

The next step will be to push to heroku and add filtering and searching.