Completed base functionality.
Added filter text area in html, filterIdeas function, and updated gitIdeas function with spans.
There are no issues with filter, with the exception of when an idea quality is upped or downed user
can see the page reload. Update functions and controller would need to be updated to only update portions
of ideas that are being updated, instead of reloading page.

The are still issues with the modal not closing and tests still need to be added.